+++ src/promptflow/tests/sdk_cli_azure_test/e2etests/test_flow_operations.py
@@ -10,6 +10,7 @@ import pytest
 from promptflow._sdk._constants import FlowType
 from promptflow._sdk._errors import FlowOperationError
 from promptflow.azure import PFClient
+from promptflow.azure._entities._flow import Flow
 
 from .._azure_utils import DEFAULT_TEST_TIMEOUT, PYTEST_TIMEOUT_METHOD
 from ..recording_utilities import is_live
@@ -17,6 +18,25 @@ from ..recording_utilities import is_live
 tests_root_dir = Path(__file__).parent.parent.parent
 flow_test_dir = tests_root_dir / "test_configs/flows"
 data_dir = tests_root_dir / "test_configs/datas"
+
+
+def create_flow(pf: PFClient, display_name) -> Flow:
+    flow_display_name = display_name
+    flow_source = flow_test_dir / "simple_fetch_url/"
+    description = "test flow"
+    tags = {"owner": "sdk"}
+    result = pf.flows.create_or_update(
+        flow=flow_source, display_name=flow_display_name, type=FlowType.STANDARD, description=description, tags=tags
+    )
+    remote_flow_dag_path = result.path
+
+    # make sure the flow is created successfully
+    assert pf.flows._storage_client._check_file_share_file_exist(remote_flow_dag_path) is True
+    assert result.display_name == flow_display_name
+    assert result.type == FlowType.STANDARD
+    assert result.tags == tags
+    assert result.path.endswith(f"/promptflow/{flow_display_name}/flow.dag.yaml")
+    return result
 
 
 @pytest.mark.timeout(timeout=DEFAULT_TEST_TIMEOUT, method=PYTEST_TIMEOUT_METHOD)
@@ -29,20 +49,17 @@ data_dir = tests_root_dir / "test_configs/datas"
 class TestFlow:
     def test_create_flow(self, pf: PFClient, randstr: Callable[[str], str]):
         flow_display_name = randstr("flow_display_name")
-        flow_source = flow_test_dir / "simple_fetch_url/"
-        description = "test flow"
-        tags = {"owner": "sdk"}
-        result = pf.flows.create_or_update(
-            flow=flow_source, display_name=flow_display_name, type=FlowType.STANDARD, description=description, tags=tags
-        )
-        remote_flow_dag_path = result.path
+        create_flow(pf, flow_display_name)
 
-        # make sure the flow is created successfully
-        assert pf.flows._storage_client._check_file_share_file_exist(remote_flow_dag_path) is True
-        assert result.display_name == flow_display_name
-        assert result.type == FlowType.STANDARD
-        assert result.tags == tags
-        assert result.path.endswith(f"/promptflow/{flow_display_name}/flow.dag.yaml")
+    def test_get_flow(self, pf: PFClient, randstr: Callable[[str], str]):
+        flow_display_name = randstr("flow_display_name")
+        flow = create_flow(pf, flow_display_name)
+        result = pf.flows.get(name=flow.name)
+
+        # assert created flow is the same as the one retrieved
+        attributes = vars(result)
+        for attr in attributes:
+            assert getattr(result, attr) == getattr(flow, attr), f"Assertion failed for attribute: {attr!r}"
 
     @pytest.mark.skipif(
         condition=not is_live(),