+++ src/promptflow/tests/test_configs/flows/python_tool_with_image_nested_api_calls/passthrough.py
@@ -1,0 +1,8 @@
+from promptflow import tool
+
+
+@tool
+def passthrough(image, call_passthrough: bool = True):
+    if call_passthrough:
+        image = passthrough(image, False)
+    return image